feat: cpu registers
Create RegisterState to represent 8 8-bit registers the CPU has access
to. Add Register and WideRegister enums that represent 8-bit addresses
used in CPU operations and the getters and setters of the RegisterState.

WideRegister addresses act as 16-bit pseudo-registers over 2 of the
8-bit ones.

AnyRegister can be used as an abstraction for any operations that do not
require a specific register width.